LayoutTests/fast/dom/HTMLElement/set-inherit-parent-true.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>
<div id="div" contenteditable="true">
    <p id="p" contenteditable="false"></p>
</div>
<script>
description('When contentEditable is set to "inherit", the attribute must be removed and element.contentEditable should return "inherit".')

document.getElementById("p").contentEditable = "inherit";

shouldBe('document.getElementById("p").hasAttribute("contentEditable")','false');
shouldBe('document.getElementById("p").contentEditable', '"inherit"');
shouldBe('document.getElementById("p").isContentEditable', 'true');
debug("FIXME: setContentEditable with true/false/inherit string is not working properly. https://bugs.webkit.org/show_bug.cgi?id=52058");
shouldBe('window.getComputedStyle(p, "").getPropertyValue("-webkit-user-modify")', '"read-write"');

document.getElementById("div").style.display= 'none';
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
